refactor(voice): extract websocket protocol types
